backend/requirements.txt
Flask
BeautifulSoup4
pandas
openpyxl
flask-cors
psycopg2-binary
flask-jwt-extended